Chat Filter System
Implements a filtering system for type chat. 
This incorporates a character replacement system that replaces what the server sees with other characters (For example, "à" would be replaced by "a"). What characters are replaced by what can be customized in `chatfilter.json`. Spaces and commas are hardcoded to be removed from the message that the server reads due to a lack of experience with .NET.
Includes a feature that replaces blocked messages with "This message has been removed." The message can be changed in `chatfilter.json`. It can be removed by making the string empty.
This system also logs messages into the console, and marks which messages are blocked.